requirements.txt
streamlit
pillow
pytesseract
streamlit-cropper
pandas
